requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
